Add a selector and shortcut bar
This idea borrowed and some of the code borrowed from YARM.

Clicking the shortcut puts a blueprint selector in your cursor that
will match just the Transport Drones entities and tiles.

Next step I want to add a handler to it that looks at the selected
items, and reports on their names/network ids, as a debugging probe.

Images for the shortcut button copied from the Shortcuts mod:
https://github.com/npc-strider/factorio-shortcuts